@@ -1,29 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using Microsoft.AspNetCore.Identity;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Text;
 
-namespace LibraryMVC.Domain.Models
+namespace LibraryMVC.Domain.Entities
 {
-    public class Account
+    public class Account 
     {
+        [Key]
         public int ID { get; set; }
-
-        [Required(ErrorMessage = "Wprowadź nickname dla swojego konta")]
-        [StringLength(30, MinimumLength = 3)]
-        public string NickName { get; set; }
+        [Required(ErrorMessage = "Wprowadź nazwę konta")]
+        public string UserName { get; set; }
+        public string NormalizedUserName { get; set; }
 
         [Required(ErrorMessage = "Wprowadź hasło")]
-        public string Password { get; set; }
+        public string PasswordHash { get; set; }
 
         [Required(ErrorMessage = "Powtórz hasło")]
         [NotMapped]
-        [Compare("Password")]
+        [Compare("PasswordHash")]
         public string ConfirmPassword { get; set; }
 
-        [Required]
-        [EmailAddress]
         public string Email { get; set; }
+        public string Roles { get; set; }
     }
 }
