@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Diagnostics;
 using LibraryMVC.Domain.Repositories;
 using Microsoft.AspNetCore.Mvc;
-using LibraryMVC.Models;
+using LibraryMVC.Entities;
 using Microsoft.AspNetCore.Authorization;
 
 namespace LibraryMVC.Controllers
