@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
-using LibraryMVC.Domain.Models;
+using LibraryMVC.Domain.Entities;
 using LibraryMVC.Domain.Repositories;
 using Microsoft.AspNetCore.Mvc;
 
