@@ -1,49 +1,54 @@
 ﻿// <auto-generated />
+using System;
 using LibraryMVC.Domain;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Migrations;
-using Microsoft.EntityFrameworkCore.Storage;
-using Microsoft.EntityFrameworkCore.Storage.Internal;
-using System;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 namespace LibraryMVC.Migrations
 {
     [DbContext(typeof(LibraryDbContext))]
-    [Migration("20190519175246_UserRegistrationAndLogin")]
-    partial class UserRegistrationAndLogin
+    [Migration("20191006190552_IdentityAccount")]
+    partial class IdentityAccount
     {
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.0.0-rtm-26452")
+                .HasAnnotation("ProductVersion", "2.2.4-servicing-10062")
+                .HasAnnotation("Relational:MaxIdentifierLength", 128)
                 .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-            modelBuilder.Entity("LibraryMVC.Domain.Models.Account", b =>
+            modelBuilder.Entity("LibraryMVC.Domain.Entities.Account", b =>
                 {
                     b.Property<int>("ID")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<string>("Email")
+                    b.Property<string>("Email");
+
+                    b.Property<string>("NormalizedUserName");
+
+                    b.Property<string>("PasswordHash")
                         .IsRequired();
 
-                    b.Property<string>("NickName")
-                        .IsRequired()
-                        .HasMaxLength(30);
+                    b.Property<string>("Roles");
 
-                    b.Property<string>("Password");
+                    b.Property<string>("UserName")
+                        .IsRequired();
 
                     b.HasKey("ID");
 
                     b.ToTable("Accounts");
                 });
 
-            modelBuilder.Entity("LibraryMVC.Domain.Models.Book", b =>
+            modelBuilder.Entity("LibraryMVC.Domain.Entities.Book", b =>
                 {
                     b.Property<int>("ID")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Author");
 
@@ -60,10 +65,11 @@ namespace LibraryMVC.Migrations
                     b.ToTable("Books");
                 });
 
-            modelBuilder.Entity("LibraryMVC.Domain.Models.BorrowedBook", b =>
+            modelBuilder.Entity("LibraryMVC.Domain.Entities.BorrowedBook", b =>
                 {
                     b.Property<int>("ID")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<int>("BookId");
 
@@ -82,10 +88,11 @@ namespace LibraryMVC.Migrations
                     b.ToTable("BorrowedBooks");
                 });
 
-            modelBuilder.Entity("LibraryMVC.Domain.Models.User", b =>
+            modelBuilder.Entity("LibraryMVC.Domain.Entities.User", b =>
                 {
                     b.Property<int>("ID")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("FirstName");
 
@@ -98,14 +105,29 @@ namespace LibraryMVC.Migrations
                     b.ToTable("Users");
                 });
 
-            modelBuilder.Entity("LibraryMVC.Domain.Models.BorrowedBook", b =>
+            modelBuilder.Entity("LibraryMVC.Domain.Entities.UserRole", b =>
                 {
-                    b.HasOne("LibraryMVC.Domain.Models.Book", "Book")
+                    b.Property<int>("ID")
+                        .ValueGeneratedOnAdd()
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
+
+                    b.Property<int>("RoleID");
+
+                    b.Property<int>("UserID");
+
+                    b.HasKey("ID");
+
+                    b.ToTable("UserRoles");
+                });
+
+            modelBuilder.Entity("LibraryMVC.Domain.Entities.BorrowedBook", b =>
+                {
+                    b.HasOne("LibraryMVC.Domain.Entities.Book", "Book")
                         .WithMany("BorrowingUsers")
                         .HasForeignKey("BookId")
                         .OnDelete(DeleteBehavior.Cascade);
 
-                    b.HasOne("LibraryMVC.Domain.Models.User", "User")
+                    b.HasOne("LibraryMVC.Domain.Entities.User", "User")
                         .WithMany("BorrowedBooks")
                         .HasForeignKey("UserId")
                         .OnDelete(DeleteBehavior.Cascade);
