@@ -1,21 +1,16 @@
 ﻿@model LibraryMVC.Domain.Entities.Account
 @{
-    ViewData["Title"] = "Register";
-
+    ViewData["Title"] = "Login";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
-<h2>Register</h2>
-
+<h2>Login</h2>
 @using (Html.BeginForm(FormMethod.Post))
 {
     <div class="row">
         <div class="col-md-4">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-                @Html.LabelFor(m => m.UserName, htmlAttributes: new { @class = "control-label" })
-                @Html.TextBoxFor(m => m.UserName, new { @class = "form-control" })
-                @Html.ValidationMessageFor(m => m.UserName, "", new { @class = "text-danger" })
-            </div>
+
             <div class="form-group">
                 @Html.LabelFor(m => m.Email, htmlAttributes: new { @class = "control-label" })
                 @Html.TextBoxFor(m => m.Email, new { @class = "form-control" })
@@ -25,11 +20,6 @@
                 @Html.LabelFor(m => m.PasswordHash, htmlAttributes: new { @class = "control-label" })
                 @Html.PasswordFor(m => m.PasswordHash, new { @class = "form-control" })
                 @Html.ValidationMessageFor(m => m.PasswordHash, "", new { @class = "text-danger" })
-            </div>
-            <div class="form-group">
-                @Html.LabelFor(m => m.ConfirmPassword, htmlAttributes: new { @class = "control-label" })
-                @Html.PasswordFor(m => m.ConfirmPassword, new { @class = "form-control" })
-                @Html.ValidationMessageFor(m => m.ConfirmPassword, "", new { @class = "text-danger" })
             </div>
 
             <div class="form-group">
