@@ -1,4 +1,4 @@
-﻿@model  LibraryMVC.Domain.Models.Book
+﻿@model  LibraryMVC.Domain.Entities.Book
 @{
     ViewData["Title"] = "Delete";
     Layout = "~/Views/Shared/_Layout.cshtml";
