@@ -1,4 +1,4 @@
-﻿@model LibraryMVC.Domain.Models.Book
+﻿@model LibraryMVC.Domain.Entities.Book
 @using LibraryMVC
 @{
     ViewData["Title"] = @Model.Name;
