@@ -1,4 +1,4 @@
-﻿@model IEnumerable<LibraryMVC.Domain.Models.Book>
+﻿@model IEnumerable<LibraryMVC.Domain.Entities.Book>
 @{
     ViewData["Title"] = "List";
 
