@@ -1,4 +1,4 @@
-﻿@model LibraryMVC.Domain.Models.Book
+﻿@model LibraryMVC.Domain.Entities.Book
 @{
     ViewData["Title"] = "Zwróć książkę";
     Layout = "~/Views/Shared/_Layout.cshtml";
