@@ -1,8 +1,8 @@
-﻿@model LibraryMVC.Domain.Models.User
+﻿@model LibraryMVC.Domain.Entities.User
 @{
     ViewData["Title"] = "Aktywacja użytkownika";
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
-<h2>Użytkownik @Model.NickName został aktywowany!</h2>
+<h2>Użytkownik został aktywowany!</h2>
 
