@@ -1,4 +1,4 @@
-﻿@model LibraryMVC.Domain.Models.User
+﻿@model LibraryMVC.Domain.Entities.User
 @{
     ViewData["Title"] = "Edit";
     Layout = "~/Views/Shared/_Layout.cshtml";
