@@ -5,7 +5,7 @@
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
-<h2>@Model.User.NickName</h2>
+
 <dt>
     Imię:
 </dt>
