@@ -1,4 +1,4 @@
-﻿@model IEnumerable<LibraryMVC.Domain.Models.User>
+﻿@model IEnumerable<LibraryMVC.Domain.Entities.User>
 @{
     ViewData["Title"] = "Activation";
     Layout = "~/Views/Shared/_Layout.cshtml";
