@@ -1,3 +1,3 @@
 @using LibraryMVC
-@using LibraryMVC.Models
+@using LibraryMVC.Entities
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
